+++ src/components/filmCardsContainer.js
@@ -1,0 +1,12 @@
+const createFilmCardsContainer = () => {
+  return (
+    `<section class="films">
+      <section class="films-list">
+        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+        <div class="films-list__container"></div>
+      </section>
+    </section>`
+  );
+};
+
+export {createFilmCardsContainer};
