+++ src/components/mostCommentedFilmsContainer.js
@@ -1,0 +1,10 @@
+const createMostCommentedFilmsContainer = () => {
+  return (
+    `<section class="films-list--extra films-list--most-commented">
+      <h2 class="films-list__title">Most commented</h2>
+      <div class="films-list__container"></div>
+    <section>`
+  );
+};
+
+export {createMostCommentedFilmsContainer};
