+++ src/components/showMoreBtn.js
@@ -1,0 +1,7 @@
+const createShowMoreBtnTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createShowMoreBtnTemplate};
