+++ src/main.js
@@ -1,145 +1,15 @@
-'use strict';
+// импорт компонентов
+import {createUserRankTemplate} from './components/userRank.js';
+import {createMenuTemplate} from './components/menu.js';
+import {createFilmCardTemplate} from './components/filmCard.js';
+import {createShowMoreBtnTemplate} from './components/showMoreBtn.js';
+import {createFilmCardsContainer} from './components/filmCardsContainer.js';
+import {createTopRatedFilmsContainer} from './components/topRatedFilmsContainer.js';
+import {createMostCommentedFilmsContainer} from './components/mostCommentedFilmsContainer';
 
+// константы
 const FILM_COUNT = 5;
 const EXTRA_FILMS_COUNT = 2;
-
-// функции, возвращающие разметку
-const createUserRankTemplate = () => {
-  return (
-    `<section class="header__profile profile">
-      <p class="profile__rating">Movie Buff</p>
-      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-    </section>`
-  );
-};
-
-const createMenuTemplate = () => {
-  return (
-    `<nav class="main-navigation">
-      <div class="main-navigation__items">
-        <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-        <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">13</span></a>
-        <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">4</span></a>
-        <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
-      </div>
-      <a href="#stats" class="main-navigation__additional">Stats</a>
-    </nav>
-
-    <ul class="sort">
-      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-      <li><a href="#" class="sort__button">Sort by date</a></li>
-      <li><a href="#" class="sort__button">Sort by rating</a></li>
-    </ul>`
-  );
-};
-
-const createFilmCardTemplate = () => {
-  return (
-    `<article class="film-card">
-      <h3 class="film-card__title">The Dance of Life</h3>
-      <p class="film-card__rating">8.3</p>
-      <p class="film-card__info">
-        <span class="film-card__year">1929</span>
-        <span class="film-card__duration">1h 55m</span>
-        <span class="film-card__genre">Musical</span>
-      </p>
-      <img src="./images/posters/the-dance-of-life.jpg" alt="" class="film-card__poster">
-      <p class="film-card__description">Burlesque comic Ralph "Skid" Johnson (Skelly), and specialty dancer Bonny Lee King (Carroll), end up together on a cold, rainy night at a tr…</p>
-      <a class="film-card__comments">5 comments</a>
-      <form class="film-card__controls">
-        <button class="film-card__controls-item button film-card__controls-item--add-to-watchlist">Add to watchlist</button>
-        <button class="film-card__controls-item button film-card__controls-item--mark-as-watched">Mark as watched</button>
-        <button class="film-card__controls-item button film-card__controls-item--favorite">Mark as favorite</button>
-      </form>
-    </article>`
-  );
-};
-
-const createShowMoreBtnTemplate = () => {
-  return (
-    `<button class="films-list__show-more">Show more</button>`
-  );
-};
-
-
-const createFilmInfoTemplate = () => {
-  return (
-    `<div class="film-details__close">
-        <button class="film-details__close-btn" type="button">close</button>
-      </div>
-      <div class="film-details__info-wrap">
-        <div class="film-details__poster">
-          <img class="film-details__poster-img" src="./images/posters/the-great-flamarion.jpg" alt="">
-
-          <p class="film-details__age">18+</p>
-        </div>
-
-        <div class="film-details__info">
-          <div class="film-details__info-head">
-            <div class="film-details__title-wrap">
-              <h3 class="film-details__title">The Great Flamarion</h3>
-              <p class="film-details__title-original">Original: The Great Flamarion</p>
-            </div>
-
-            <div class="film-details__rating">
-              <p class="film-details__total-rating">8.9</p>
-            </div>
-          </div>
-
-          <table class="film-details__table">
-            <tr class="film-details__row">
-              <td class="film-details__term">Director</td>
-              <td class="film-details__cell">Anthony Mann</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Writers</td>
-              <td class="film-details__cell">Anne Wigton, Heinz Herald, Richard Weil</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Actors</td>
-              <td class="film-details__cell">Erich von Stroheim, Mary Beth Hughes, Dan Duryea</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Release Date</td>
-              <td class="film-details__cell">30 March 1945</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Runtime</td>
-              <td class="film-details__cell">1h 18m</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Country</td>
-              <td class="film-details__cell">USA</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Genres</td>
-              <td class="film-details__cell">
-                <span class="film-details__genre">Drama</span>
-                <span class="film-details__genre">Film-Noir</span>
-                <span class="film-details__genre">Mystery</span></td>
-            </tr>
-          </table>
-
-          <p class="film-details__film-description">
-            The film opens following a murder at a cabaret in Mexico City in 1936, and then presents the events leading up to it in flashback. The Great Flamarion (Erich von Stroheim) is an arrogant, friendless, and misogynous marksman who displays his trick gunshot act in the vaudeville circuit. His show features a beautiful assistant, Connie (Mary Beth Hughes) and her drunken husband Al (Dan Duryea), Flamarion's other assistant. Flamarion falls in love with Connie, the movie's femme fatale, and is soon manipulated by her into killing her no good husband during one of their acts.
-          </p>
-        </div>
-      </div>
-
-    <section class="film-details__controls">
-      <input type="checkbox" class="film-details__control-input visually-hidden" id="watchlist" name="watchlist">
-      <label for="watchlist" class="film-details__control-label film-details__control-label--watchlist">Add to watchlist</label>
-
-      <input type="checkbox" class="film-details__control-input visually-hidden" id="watched" name="watched">
-      <label for="watched" class="film-details__control-label film-details__control-label--watched">Already watched</label>
-
-      <input type="checkbox" class="film-details__control-input visually-hidden" id="favorite" name="favorite">
-      <label for="favorite" class="film-details__control-label film-details__control-label--favorite">Add to favorites</label>
-    </section>`
-  );
-};
-createFilmInfoTemplate(); // использование переменной для линтера
-
 
 // рендеринг
 const renderTemplate = (container, template, place) => {
@@ -154,24 +24,10 @@ renderTemplate(containerUserRank, createUserRankTemplate(), `beforeend`);
 const containerMain = document.querySelector(`.main`);
 renderTemplate(containerMain, createMenuTemplate(), `afterbegin`);
 
-// карточки фильма
-// создание контейнера для карточек фильмов
-const renderFilmCardsContainer = () => {
-  return (
-    `<section class="films">
-      <section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-        <div class="films-list__container"></div>
-      </section>
-    </section>`
-  );
-};
-
 // рендер контейнера для карточек фильмов
-renderTemplate(containerMain, renderFilmCardsContainer(), `beforeend`);
+renderTemplate(containerMain, createFilmCardsContainer(), `beforeend`);
 
 // рендер карточек фильмов
-
 const filmCardContainer = document.querySelector(`.films-list__container`);
 
 for (let i = 0; i < FILM_COUNT; i++) {
@@ -181,36 +37,15 @@ for (let i = 0; i < FILM_COUNT; i++) {
 // рендер кнопки
 renderTemplate(filmCardContainer, createShowMoreBtnTemplate(), `afterend`);
 
-// создание контейнеров top-rated- и most-commented-фильмов
-const createTopRatedFilmsContainer = () => {
-  return (
-    `<section class="films-list--extra">
-      <h2 class="films-list__title">Top rated</h2>
-      <div class="films-list__container"></div>
-    <section>`
-  );
-};
-
-const createMostCommentedFilmsContainer = () => {
-  return (
-    `<section class="films-list--extra films-list--most-commented">
-      <h2 class="films-list__title">Most commented</h2>
-      <div class="films-list__container"></div>
-    <section>`
-  );
-};
-
 // рендер mostCommented-контейнера
 renderTemplate(document.querySelector(`.films-list`), createMostCommentedFilmsContainer(), `afterend`);
 
 // рендер карточек mostCommented-фильмов
-
 const mostCommentedFilmsContainer = document.querySelector(`.films-list--most-commented .films-list__container`);
 
 for (let i = 0; i < EXTRA_FILMS_COUNT; i++) {
   renderTemplate(mostCommentedFilmsContainer, createFilmCardTemplate(), `afterbegin`);
 }
-
 
 // рендер topRated-контейнера
 renderTemplate(document.querySelector(`.films-list`), createTopRatedFilmsContainer(), `afterend`);
